requirements.txt
appdirs==1.4.4
art==5.1
autopep8==1.5.3
certifi==2020.6.20
click==7.1.2
distlib==0.3.1
filelock==3.0.12
fuzzywuzzy==0.18.0
gensim==4.0.0
grpcio==1.36.1
grpcio-tools==1.36.1
joblib==1.0.1
nltk==3.5
numpy==1.20.2
protobuf==3.15.6
pycodestyle==2.6.0
python-slugify==4.0.1
regex==2021.3.17
scipy==1.6.2
six==1.15.0
slugify==0.0.1
smart-open==5.0.0
text-unidecode==1.3
textblob==0.15.3
toml==0.10.1
tqdm==4.59.0
wincertstore==0.2
zipp==3.1.0
